+++ test/helper-word-test.js
@@ -9,7 +9,7 @@ var options = {}
 describe('Parser Helpers', function () {
   describe('helpers.word()', function () {
     describe('with a whole word', function () {
-      var Parser = atokParser.createParser('./wordHelperParser.js', 'options')
+      var Parser = atokParser.createParser('./parsers/wordHelperParser.js', 'options')
       var p = new Parser(options)
 
       it('should parse it', function (done) {
@@ -31,7 +31,7 @@ describe('Parser Helpers', function () {
     })
 
     describe('with a split up word', function () {
-      var Parser = atokParser.createParser('./wordHelperParser.js', 'options')
+      var Parser = atokParser.createParser('./parsers/wordHelperParser.js', 'options')
       var p = new Parser(options)
 
       it('should parse it', function (done) {
@@ -55,7 +55,7 @@ describe('Parser Helpers', function () {
     })
 
     describe('with a non ending word', function () {
-      var Parser = atokParser.createParser('./wordHelperParser.js', 'options')
+      var Parser = atokParser.createParser('./parsers/wordHelperParser.js', 'options')
       var p = new Parser(options)
       
       it('should not parse it', function (done) {